adding more unit-tests
